Move `std::io::Read` to `alloc::io`
