+++ library/alloc/src/lib.rs
@@ -121,6 +121,7 @@
 #![feature(core_io)]
 #![feature(core_io_borrowed_buf)]
 #![feature(core_io_internals)]
+#![feature(cursor_split)]
 #![feature(deprecated_suggestion)]
 #![feature(deref_pure_trait)]
 #![feature(diagnostic_on_move)]
@@ -140,6 +141,7 @@
 #![feature(inplace_iteration)]
 #![feature(io_const_error)]
 #![feature(io_const_error_internals)]
+#![feature(io_slice_as_bytes)]
 #![feature(iter_advance_by)]
 #![feature(iter_next_chunk)]
 #![feature(layout_for_ptr)]
@@ -147,6 +149,7 @@
 #![feature(likely_unlikely)]
 #![feature(local_waker)]
 #![feature(maybe_uninit_array_assume_init)]
+#![feature(maybe_uninit_fill)]
 #![feature(maybe_uninit_uninit_array_transpose)]
 #![feature(panic_internals)]
 #![feature(pattern)]
@@ -218,6 +221,7 @@
 //
 // Rustdoc features:
 #![feature(doc_cfg)]
+#![feature(doc_notable_trait)]
 // Technically, this is a bug in rustdoc: rustdoc sees the documentation on `#[lang = slice_alloc]`
 // blocks is for `&[T]`, which also has documentation using this feature in `core`, and gets mad
 // that the feature-gate isn't enabled. Ideally, it wouldn't check for the feature gate for docs
